ENH: Add h5 script
AnalyzeH5.py is a refactoring of the analysis in SimpleClustersParallelSlice.
The idea is to have analyses that run on one or more h5 files independent of the data source (psana1 or 2 or rogue).

Can add more analyses to AnalyzeH5, concept is to label the h5 with the type
of analysis (cluster, linearity, ...) and have AnalyzeH5 take it from there.